+++ exe-common/src/lib.rs
@@ -16,6 +16,7 @@ extern crate hyper;
 extern crate tokio_core;
 
 mod mstream;
+pub mod utils;
 pub mod network;
 pub mod config;
 pub mod sync;
