+++ exe-common/src/network/hermes.rs
@@ -1,4 +1,4 @@
-use cardano::block::{BlockHeader, Block, HeaderHash};
+use cardano::block::{block, BlockHeader, Block, HeaderHash};
 use storage::{self, Storage, tmpfile::{TmpFile}};
 use std::io::{Write, Seek, SeekFrom};
 use std::time::{SystemTime};
